Clean up character offset and advance.

+++ code/Alpha/Rendering/Font/Font.h
@@ -28,10 +28,10 @@ public:
     uint32_t m_width; // width of character in pixels
     uint32_t m_height; // height of character in pixels
 
-    // these need to be normalized based on size of font
-    float m_xoffset; // offset from current cursor pos to left side of character
-    float m_yoffset; // offset from top of line to top of character
-    float m_xadvance; // how far to move to right for next character
+    //all those values are in pixels.
+    int32_t m_xoffset; // offset from current cursor pos to left side of character
+    int32_t m_yoffset; // offset from top of line to top of character
+    int32_t m_xadvance; // how far to move to right for next character
 };
 
 class FontKerning
